moved both class and course Edit/Creat to individual file(.jsx)
