+++ admin/src/courses.jsx
@@ -1,4 +1,4 @@
-import { Datagrid, List, TextField, BooleanField, EditButton } from 'react-admin';
+import { Datagrid, List, TextField, BooleanField, EditButton, Edit, Create, TextInput, BooleanInput, SimpleForm } from 'react-admin';
 
 
 export const CourseList = () => (
@@ -18,4 +18,40 @@ export const CourseList = () => (
             <EditButton />
         </Datagrid>
     </List>
+);
+// ?? Need to use <TextInput> instaead of TextField for Edit/ Create ?? 
+export const CourseEdit = () => (
+    <Edit title="Edit courses">
+        <SimpleForm>
+            <TextField source="id" />
+            <TextField source="name" />
+            <TextField source="shortname" />
+            <TextField source="class" />
+            <TextField source="points" />
+            <TextField source="startDate" />
+            <TextField source="endDate" />
+            <TextField source="plan" />
+            <TextField source="invoiceItem" />
+            <TextField source="hoursPerDay" />
+            <BooleanField source="hide" />
+        </SimpleForm>
+    </Edit>
+);
+
+export const CourseCreate = props => (
+    <Create {...props}>
+        <SimpleForm>
+            <TextField source="id" />
+            <TextField source="name" />
+            <TextField source="shortname" />
+            <TextField source="class" />
+            <TextField source="points" />
+            <TextField source="startDate" />
+            <TextField source="endDate" />
+            <TextField source="plan" />
+            <TextField source="invoiceItem" />
+            <TextField source="hoursPerDay" />
+            <BooleanField source="hide" />
+        </SimpleForm>
+    </Create>
 );
