+++ admin/src/ivoice-item.jsx
@@ -1,4 +1,4 @@
-import { Datagrid, List, TextField, EditButton, Edit, SimpleForm, TextInput, Create, SimpleFormIteratorItem,  } from 'react-admin';
+import { Datagrid, List, TextField, EditButton, Edit, SimpleForm, TextInput, Create, SimpleForm,  } from 'react-admin';
 
 export const InvoiceItemList = () => (
     <List>
